+++ src/BT/ConcurrentNode.hpp
@@ -19,8 +19,8 @@ public:
     virtual Ptr getCopy() override;
 
 private:
-    virtual State::StateType performAction() override;
-    virtual State::StateType continueAction() override;
+    virtual StateType performAction() override;
+    virtual StateType continueAction() override;
 
 };