+++ src/UnitTest/TestLuaFactory.cpp
@@ -81,13 +81,13 @@ TEST(BehaviorLuaFactory, Factory)
         BehaviorNode::Ptr tree = blf.createTreeFromFile("TestLuaFactoryScript0.lua");
         ASSERT_TRUE(tree);
 
-        EXPECT_EQ(tree->activate(), BehaviorNode::State::READY_SUCCESS);
+        EXPECT_EQ(tree->activate(), BehaviorNode::StateType::BT_READY_SUCCESS);
 
         EXPECT_EQ(1, a0);
         EXPECT_EQ(1, a1);
         EXPECT_EQ(1, a2);
 
-        EXPECT_EQ(tree->activate(), BehaviorNode::State::READY_SUCCESS);
+        EXPECT_EQ(tree->activate(), BehaviorNode::StateType::BT_READY_SUCCESS);
 
         EXPECT_EQ(2, a0);
         EXPECT_EQ(2, a1);
@@ -99,7 +99,7 @@ TEST(BehaviorLuaFactory, Factory)
         BehaviorNode::Ptr tree = blf.createTreeFromFile("TestLuaFactoryScript1.lua");
         ASSERT_TRUE(tree);
 
-        EXPECT_EQ(tree->activate(), BehaviorNode::State::READY_SUCCESS);
+        EXPECT_EQ(tree->activate(), BehaviorNode::StateType::BT_READY_SUCCESS);
 
         EXPECT_EQ(1, a0);
     }
@@ -117,7 +117,7 @@ TEST(BehaviorLuaFactory, Factory)
         BehaviorNode::Ptr tree = blf.createTreeFromFile("TestLuaFactoryScript2.lua");
         ASSERT_TRUE(tree);
 
-        EXPECT_EQ(tree->activate(), BehaviorNode::State::READY_SUCCESS);
+        EXPECT_EQ(tree->activate(), BehaviorNode::StateType::BT_READY_SUCCESS);
 
         EXPECT_EQ(1, a.state);
     }
@@ -140,7 +140,7 @@ TEST(BehaviorLuaFactory, FactoryCustom)
         BehaviorNode::Ptr tree = blf.createTreeFromFile("TestLuaFactoryScript3.lua");
         ASSERT_TRUE(tree);
 
-        EXPECT_EQ(tree->activate(), BehaviorNode::State::READY_SUCCESS);
+        EXPECT_EQ(tree->activate(), BehaviorNode::StateType::BT_READY_SUCCESS);
 
         EXPECT_EQ(1, a0);
         EXPECT_EQ(1, a1);
@@ -265,7 +265,7 @@ TEST(BehaviorLuaFactory, ExposeLib)
         "";
 
     auto tree = blf.createTreeFromScript(luaScript, strlen(luaScript));
-    EXPECT_EQ(tree->activate(), BehaviorNode::State::READY_SUCCESS);
+    EXPECT_EQ(tree->activate(), BehaviorNode::StateType::BT_READY_SUCCESS);
 }
 
 TEST(BehaviorLuaFactory, getLuaStatesInTree)